requirements.txt
rimdev-lib==0.0.14
fastapi==0.78.0
uvicorn==0.17.6
PyMySQL==1.0.2
passlib==1.7.4
python-jose==3.3.0
bcrypt==3.2.2
python_multipart==0.0.5
python_dotenv==0.20.0
SQLAlchemy==1.4.37
pandas==1.4.2
bs4==0.0.1
lxml==4.9.0
requests==2.28.0
krcg==2.25